fix rename the test file
